fix(bin): symlink xterm to /usr/bin/urxvtc
so that it is possible to open gnome desktop entries which are
to be run in a terminal.
This issue arises as glib has a number of hardcoded terminals to
use when opening terminal desktop entries and rxvt-unicode is not
in this list (see https://gitlab.gnome.org/GNOME/glib/issues/338).

The only other option would be to override all terminal run apps
with user specific desktop entries which have "Terminal=false" and
instead run urxvtc directly e.g. "Exec=/usr/bin/urxvtc -e htop".

As it does not seem like there will be any solution on the glib
side anytime soon and even distributions have implemented their
own workarounds (https://github.com/mate-desktop/mate-panel/issues/57),
this is here to stay