feat(bin): add script to start alpine container
by using systemd's nspawn command. The container can be used for quick
testing and is possible to manage later with systemd's machinectl.

This is a lot easier than setting up a docker container, does not
require any dependencies and is good for a throwaway. Also does it setup
a full system containing all of lib/bin as normally.